Improved Controllers, Changed Payment Schema , Moved Payment Routes to User, Moved Razorpay Service to Config.
