+++ backend/models/User/paymentModel.js
@@ -1,0 +1,53 @@
+const mongoose = require("mongoose");
+
+const paymentSchema = new mongoose.Schema(
+  {
+
+      orderId: { 
+        type: String,
+        required: [true, "Order ID is required"]
+      },
+      paymentId: { 
+        type: String 
+      },
+      amount: { 
+        type: Number,
+        required: [true, "Amount is required"]
+      },
+      status: {
+        type: String,
+        enum: ["completed", "pending", "cancelled"],
+        default: "pending",
+      },
+      clientName: { 
+        type: String,
+        required: [true, "Client name is required"]
+      },
+      clientId: { 
+        type: String,
+        required: [true, "Client ID is required"]
+      },
+      therapistId: { 
+        type: String,
+        required: [true, "Therapist ID is required"]
+      },
+      therapistName: { 
+        type: String,
+        required: [true, "Therapist Name is required"]
+      },
+      appointmentType: { 
+        type: String,
+        required: [true, "Appointment type is required"]
+      },
+      therapyType: { 
+        type: String,
+        required: [true, "Therapy type is required"]
+      },
+      createdAt: { 
+          type: Date,
+          default: Date.now 
+      },
+  });
+
+const Payment = mongoose.model("Payment", paymentSchema);
+module.exports = { Payment , mongoose };
