+++ backend/routes/userRoute.js
@@ -4,6 +4,8 @@ const {registerUser, loginUser, logout, getUser,upcomingGroups}=require("../cont
 const { protect } = require("../middleware/authMiddleware");
 const { getGroupTherapySessions, createGroupTherapySession } = require("../controllers/User/groupTherapy");
 const { getSupportGroupSessions, createSupportGroupSession } = require("../controllers/User/supportgroup");
+const {newOrder, verifyPayment } = require("../controllers/User/paymentController/orders")
+const { getAllPayments, getPaymentById, getOrderById, getbyClientId, getByTherapistId } = require("../controllers/User/paymentController/paymentDetails");
 
 
 router.get("/Grouptherapy", getGroupTherapySessions);           
@@ -13,6 +15,16 @@ router.get("/Supportgroup", getSupportGroupSessions);
 router.post("/Supportgroup", createSupportGroupSession);
 
 
+router.post('/newOrder', newOrder);
+router.post('/verifyPayment', verifyPayment);
+
+router.get('/getAllPayments', getAllPayments);
+router.get('/getPaymentById/:id',getPaymentById);
+router.get('/getOrderById/:orderId', getOrderById);
+router.get('/getbyClientId/:clientId', getbyClientId);
+router.get('/getByTherapistId/:therapistId', getByTherapistId);
+//Delete Payment Records by Order ID
+// router.delete('/deleteOrder/:orderId', deleteOrder)
 
 router.post("/register",registerUser);
 router.post("/login",loginUser)
